requirements.txt
Django==1.9.12
django-crispy-forms==1.6.1
django-enumfield==1.3b2
django-nested-inline==0.3.6
django-phonenumber-field